+++ tests/0700_pages_registry/10_add_page.py
@@ -3,7 +3,7 @@ import os
 from jinja2 import Environment as Jinja2Environment
 from jinja2 import FileSystemLoader
 
-from optimus.pages.views.base import PageTemplateView
+from optimus.pages.views import PageTemplateView
 from optimus.pages.registry import PageRegistry