+++ tests/1000_interfaces/03_build.py
@@ -4,7 +4,7 @@ import os
 from optimus.interfaces.build import builder_interface
 from optimus.interfaces.starter import starter_interface
 from optimus.logs import set_loggers_level
-from optimus.pages.views.base import PageTemplateView
+from optimus.pages.views import PageTemplateView
 
 
 class DummyView(PageTemplateView):